Variable types, string generation expanded, fixes
Added generation of texture coordinate strings for future use
Added variable types for non numpy types
Added translation method
Fixed generation of output directory if it is not existant
Renamed some Edges to Corners
Added some documentation

+++ AvObjConverter.py
@@ -3,6 +3,7 @@ import xml.etree.ElementTree
 
 import itertools
 from sys import argv
+from typing import List
 
 import numpy as np
 import sys
@@ -22,8 +23,14 @@ BLOCK = 'block'
 ITEM = 'item'
 
 
+Vertex = List[float]
+Face = List[int]
+Normal = List[float]
+UV = List[float]
+
+
 class Block:
-    def __init__(self, index, x_min, x_max, y_min, y_max, z_min, z_max, type, material, look, up, color):
+    def __init__(self, index: int, x_min: float, x_max: float, y_min: float, y_max: float, z_min: float, z_max: float, type: int, material: int, look: int, up: int, color: str):
         self.material = float(material)
         self.x_max = float(x_max)
         self.y_min = float(y_min)
@@ -48,8 +55,8 @@ class AvToObjConverter:
         if self.blocks is None:
             self.blocks = readSourceFile(inputFile)
         try:
-            while os.path.exists(outputFile) and os.path.getsize(outputFile) > 0:
-                # outputFile = outputFile[:-4] + '1' + outputFile[-4:]
+            os.makedirs(os.path.dirname(outputFile), exist_ok=True)
+            if os.path.exists(outputFile) and os.path.getsize(outputFile) > 0:
                 os.remove(outputFile)
                 print(outputFile)
 
@@ -57,16 +64,16 @@ class AvToObjConverter:
             file.write("mtllib Avorion_Materials.mtl\n")
             for block in self.blocks:
                 file.write(self.convertBlockToObj(block))
-
             file.close()
+
         except FileNotFoundError:
             print('Could not open output file')
 
     def convertBlockToObj(self, block):
         out_string = ""
         name = "Block"
-        print("Converting block with index {}, type = {}, look = {} and up = {}".format(block.index, block.type,
-                                                                                        block.look, block.up))
+        print("Converting block with index {}, type = {}, look = {} and up = {}".format(
+            block.index, block.type, block.look, block.up))
         faces = None
         # vertex_normals = None
         v_indices = []
@@ -82,13 +89,13 @@ class AvToObjConverter:
             v_indices = [1, 2, 5, 6, 7, 8]
             f_indices = [2, 4, 6, 7, 16]
 
-        elif block.type in [103, 107, 154, 174, 188, 194, 514]:  # Edge with 5 vertices
-            name = "5 Vertex Edge"
+        elif block.type in [103, 107, 154, 174, 188, 194, 514]:  # Corner with 5 vertices
+            name = "5 Vertex Corner"
             v_indices = [1, 2, 5, 6, 7]
             f_indices = [2, 7, 10, 13, 14]
 
-        elif block.type in [101, 105, 152, 172, 186, 192, 512]:  # Edge with 4 vertices
-            name = "4 Vertex Edge"
+        elif block.type in [101, 105, 152, 172, 186, 192, 512]:  # Corner with 4 vertices
+            name = "4 Vertex Corner"
             v_indices = [1, 5, 6, 7]
             f_indices = [7, 8, 10, 15]
 
@@ -109,11 +116,10 @@ class AvToObjConverter:
         # vertex_normals = rotateVectors(vertex_normals, block.look, block.up)
         if faces is None:
             faces = genfaces
-        if faces is not None:
-            out_string += self.generateObjectString(vertices, faces, vertex_normals, "{} {}".format(name, block.index))
+        out_string += self.generateObjectString(vertices, faces, vertex_normals, "{} {}".format(name, block.index))
         return out_string
 
-    def generate(self, v_indices, f_indices):
+    def generate(self, v_indices: List[int], f_indices: List[int]):
 
         vertices = [
             [1, -1, -1],  # 1 Bottom left front
@@ -187,21 +193,30 @@ class AvToObjConverter:
 
         return out_vertices, out_faces, out_normals
 
-    def generateObjectString(self, vertices, faces, vertex_normals, name, uvs=None):
+    def generateObjectString(self, vertices: List[Vertex], faces: List[Face],
+                             vertex_normals: List[Normal], name: str, uvs: List[UV] = None, smoothing: bool = False):
         out_string = "o {}\n".format(name)
-
-        for vertex in vertices:
+        smoothing = 1 if smoothing else 0
+        for vertex in vertices:  # Add the strings for the vertices
             out_string += "v {} {} {}\n".format(vertex[0], vertex[1], vertex[2])
 
-        if vertex_normals is not None:
+        if uvs is not None:  # Add UV coordinates to the object
+            for uv in uvs:
+                out_string += "vt "
+                for coord in uv:
+                    out_string += "{} ".format(coord)
+                out_string += "\n"
+
+        if vertex_normals is not None:  # Add the Strings for the vertex normals
             for normal in vertex_normals:
                 out_string += "vn "
                 for value in normal:
                     out_string += "{} ".format(value)
                 out_string += "\n"
-        out_string += "usemtl Standard\n"
-        out_string += "s off\n"
-        if faces is not None:
+
+        out_string += "usemtl Standard\n"  # Add strign for the used material
+        out_string += "s {}\n".format(smoothing)  # Add string for the enabling/disabling of smoothing
+        if faces is not None:  # Add the strings for the faces of the object
             for face in faces:
                 out_string += "f "
                 for vertex in face:
@@ -218,12 +233,12 @@ class AvToObjConverter:
         return out_string
 
 
-def normalize(vector):
+def normalize(vector: Vertex):
     length = 0
     for element in vector:
         length += element ** 2
     length = np.math.sqrt(length)
-    vector = [entry/length for entry in vector]
+    vector = [entry / length for entry in vector]
     return vector
 
 
@@ -248,13 +263,21 @@ def readSourceFile(file):
     return blocks
 
 
-def rotateVectors(vectors, look, up):
+def rotateVectors(vectors: List[Vertex], look: int, up: int):
+    """
+    Rotates the given vectors to correspond with the given values for look and up.
+
+    :param vectors: List of vectors to rotate
+    :param look: Look value to orient the vectors to (Determines the face to the back)
+    :param up: Up value to orient the vecturs to (Determines the face facing up)
+    :return: The rotated vectors
+    """
     return_vectors = []
     for vector in vectors:
         vector.append(1)
 
         temp_result = np.matrix(vector).transpose()
-        if look == 5:  # Vorne
+        if look == 5:  #
             if up == 0:
                 temp_result = rotate(temp_result, z_degrees=90)
             elif up == 2:
@@ -316,7 +339,7 @@ def rotateVectors(vectors, look, up):
     return return_vectors
 
 
-def transformBlock(verts, vert_normals, look, up, block=None):
+def transformBlock(verts: List[List[float]], vert_normals: List[List[float]], look: int, up: int, block=None):
     vertices = verts.copy()
     normals = vert_normals.copy()
     output_vertices = []
@@ -340,14 +363,15 @@ def transformBlock(verts, vert_normals, look, up, block=None):
 
         temp_result = np.matrix(vertex).transpose()
         temp_result = scale(temp_result, scale_amounts[0], scale_amounts[1], scale_amounts[2])
-        temp_result = getTranslationMatrix(translation_vector[0], translation_vector[1],
-                                           translation_vector[2]) * temp_result
+        # temp_result = getTranslationMatrix(translation_vector[0], translation_vector[1],
+        #                                    translation_vector[2]) * temp_result
+        temp_result = translate(temp_result, translation_vector[0], translation_vector[1], translation_vector[2])
 
         temp_result = temp_result.tolist()
         result = [np.round(temp_result[0][0], 15), np.round(temp_result[1][0], 15), np.round(temp_result[2][0], 15)]
         output_vertices.append(result)
 
-    for normal in normals:
+    for normal in normals:  # Update the normals to work with the scaling  -- Broken
         normal.append(0)
 
         temp_result = np.matrix(normal).transpose()
@@ -359,7 +383,12 @@ def transformBlock(verts, vert_normals, look, up, block=None):
     return output_vertices, output_normals
 
 
-def getTranslationMatrix(x, y, z):
+def translate(vector, x, y, z):
+    out_vector = _getTranslationMatrix(x, y, z).dot(vector)
+    return out_vector
+
+
+def _getTranslationMatrix(x, y, z):
     return np.matrix([[1, 0, 0, x], [0, 1, 0, y], [0, 0, 1, z], [0, 0, 0, 1]])
 
 
@@ -372,7 +401,7 @@ def scale(vector, x_amount=0, y_amount=0, z_amount=0):
     return result
 
 
-def rotate(vector, x_degrees=0, y_degrees=0, z_degrees=0):
+def rotate(vector: np.ndarray, x_degrees: float = 0, y_degrees: float = 0, z_degrees: float = 0):
     xMatrix, yMatrix, zMatrix = np.identity(4), np.identity(4), np.identity(4)
     if x_degrees != 0:
         c, s = np.cos(np.radians(x_degrees)), np.sin(np.radians(x_degrees))
@@ -407,3 +436,4 @@ conv = AvToObjConverter()
 conv.convertToObj(path + ship_name + '.xml', './out/' + ship_name + '.obj')
 
 scale(np.matrix([1, 1, 1, 0]).transpose(), 5, 6, 7)
+rotateVectors([], None, None)
